+++ src/klibc/dynarray.h
@@ -16,7 +16,7 @@ typedef struct {
     int64_t array_size;
 } dynarray_t;
 
-int dynarray_add(dynarray_t *dynarray, void *element, uint64_t size_of_elem);
+int64_t dynarray_add(dynarray_t *dynarray, void *element, uint64_t size_of_elem);
 int dynarray_remove(dynarray_t *dynarray, int64_t element);
 void *dynarray_getelem(dynarray_t *dynarray, int64_t element);
 void dynarray_unref(dynarray_t *dynarray, int64_t element);
